@@ -1,7 +1,7 @@
 package com.rentpath.motif.sample;
 
 import android.content.Context;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.Toolbar;
 import android.util.AttributeSet;
 import android.view.View;
 
